Created default handlers for most types of requests and wired them into existing API endpoints
I did this because I noticed that pretty much all of our endpoints are doing the same
core actions, and it was resulting in a lot of code being copied around, making things
very error prone, hard to refactor, and easy to break. This change cuts our API codebase to
about a third of what it was, and it makes adding new endpoints with full RESTful functionality
completely trivial; only adding custom functionality should require actual work now.